init_address will now create and return viewingkey
